build: fixing missing types folder
