+++ packages/auth/tsconfig.json
@@ -4,7 +4,6 @@
     "compilerOptions": {
         "rootDir": "lib",
         "outDir": "dist"
-
     },
 
     "exclude": [
