+++ packages/server-client/lib/types.ts
@@ -1,12 +1,1 @@
-export type APILink = {
-    href: string
-    // TODO Add more from RFC 5988?
-}
-
-export type APILinks = {
-    _links: {
-        self: APILink
-        next?: APILink
-        prev?: APILink
-    }
-}
+export * from './types/index';
