+++ packages/server-client/lib/types/APILink.ts
@@ -1,0 +1,4 @@
+export type APILink = {
+    href: string
+    // TODO Add more from RFC 5988?
+}
