+++ packages/server-client/lib/types/APILinks.ts
@@ -1,0 +1,9 @@
+import { APILink } from './APILink';
+
+export type APILinks = {
+    _links: {
+        self: APILink
+        next?: APILink
+        prev?: APILink
+    }
+}
