+++ packages/server-client/lib/types/ConfigParams.ts
@@ -1,0 +1,7 @@
+export type ConfigParams = {
+    restHost: string
+    apiHost: string
+    videoHost: string
+    responseType: string
+    timeout: number
+}
